Add files via upload
File for backtesting

Trading strategy:
We will iterate over the time period from 2016-05-30 to 2020-06-19. If any candlestick pattern is spotted, we apply the corresponding long/short strategy and profit.
On any given date, if multiple patterns are spotted, we will diversify the risk by evenly investing the principal to each stock. 
(1) for Long patterns, buy at the opening price of the day succeeding the pattern, and sell when the stock price moves upward 2% from our buying price. If the stock price moves downward more than 2% from our buying prices, we also sell the stock as a way to manage the risk. If the price stays in the +/- 2% range relative to our buying price in the considered time interval, we sell at the closing price of the last day of the considered time interval. 
(2) for Short patterns, the trading strategy is harder to implement due to the nature of short selling. For the time being, the short patterns are neglected. 

We will compare our trading strategy with the QQQ ETF on a five-year timespan, by running PnL analysis year by year and calculating variance.